One more
git-svn-id: https://www.soclib.fr/svn/trunk/soclib@2533 2cf4c1f0-a32b-0410-9068-d80f8be93c97

+++ soclib/module/connectivity_component/vci_multi_tty/tlmdt/source/src/vci_multi_tty.cpp
@@ -190,7 +190,7 @@ std::cout << "[" << name() << "] time = " << std::dec << time.value()
             term = cell / TTY_SPAN;
             payload.set_response_status(tlm::TLM_OK_RESPONSE);
 
-	        assert ( (term < (int)m_term.size() ) &&
+	        assert ( (term < m_term.size() ) &&
             " Illegal TTY access : terminal index larger than number of terminals");
 	    
             if ( extension_pointer->get_command() == VCI_READ_COMMAND )
